removed unused code and moved all classes from util to somewhere else
