+++ pe-gui/src/main/java/org/saar/gui/font/FontCharacter.kt
@@ -1,6 +1,6 @@
 package org.saar.gui.font
 
-import org.saar.utils.Box2i
+import org.saar.maths.Box2i
 
 data class FontCharacter(
     val char: Char,
