+++ pe-math/src/main/java/org/saar/maths/Box2f.kt
@@ -1,4 +1,4 @@
-package org.saar.utils
+package org.saar.maths
 
 data class Box2f(val x0: Float, val y0: Float,
                  val x1: Float, val y1: Float) {
