+++ pe-math/src/main/java/org/saar/maths/MoreDelegates.kt
@@ -1,4 +1,4 @@
-package org.saar.utils
+package org.saar.maths
 
 import kotlin.properties.ObservableProperty
 import kotlin.reflect.KProperty
