--- pe-util/src/main/java/org/saar/utils/Reference.java
@@ -1,12 +1,0 @@
-package org.saar.utils;
-
-public interface Reference<T> {
-
-    /**
-     * Returns the value
-     *
-     * @return the value
-     */
-    T get();
-
-}
