--- pe-util/src/main/java/org/saar/utils/lengths/ILength.java
@@ -1,7 +1,0 @@
-package org.saar.utils.lengths;
-
-public interface ILength {
-
-    int proportionTo(int length);
-
-}
