--- pe-util/src/main/java/org/saar/utils/lengths/Percent.java
@@ -1,21 +1,0 @@
-package org.saar.utils.lengths;
-
-public class Percent implements ILength {
-
-    public static final Percent ZERO = new Percent(0);
-
-    private final float value;
-
-    protected Percent(float value) {
-        this.value = value;
-    }
-
-    @Override
-    public int proportionTo(int length) {
-        return (int) (value * .01f * length);
-    }
-
-    public float getValue() {
-        return value;
-    }
-}
