--- pe-util/src/main/java/org/saar/utils/lengths/Pixels.java
@@ -1,19 +1,0 @@
-package org.saar.utils.lengths;
-
-public class Pixels implements ILength{
-
-    private final int value;
-
-    public Pixels(int value) {
-        this.value = value;
-    }
-
-    @Override
-    public int proportionTo(int length) {
-        return value;
-    }
-
-    public int getValue() {
-        return value;
-    }
-}
