--- pe-util/src/main/java/org/saar/utils/list/ListChangeEvent.kt
@@ -1,6 +1,0 @@
-package org.saar.utils.list
-
-data class ListChangeEvent<T>(
-        val list: ObservableList<T>,
-        val added: List<T>,
-        val removed: List<T>)
