--- pe-util/src/main/java/org/saar/utils/list/ListChangeListener.java
@@ -1,7 +1,0 @@
-package org.saar.utils.list;
-
-public interface ListChangeListener<T> {
-
-    void changed(ListChangeEvent<T> t);
-
-}
